+++ source/renderer/app/components/widgets/RadioSet.js
@@ -1,5 +1,6 @@
 // @flow
 import React, { Component } from 'react';
+import classnames from 'classnames';
 import type { Node } from 'react';
 import { Radio } from 'react-polymorph/lib/components/Radio';
 import { RadioSkin } from 'react-polymorph/lib/skins/simple/RadioSkin';
@@ -10,6 +11,7 @@ import stylesOverride from './RadioOverride.scss';
 type Props = {
   label?: Node,
   items: Array<any>,
+  verticallyAligned?: boolean,
 };
 
 type RadioProps = {
@@ -28,11 +30,15 @@ type RadioProps = {
 @observer
 export default class RadioSet extends Component<Props> {
   render() {
-    const { label, items } = this.props;
+    const { label, items, verticallyAligned } = this.props;
+    const radiosContainerStyles = classnames([
+      styles.radiosContainer,
+      verticallyAligned ? styles.verticallyAligned : null,
+    ]);
     return (
       <div className={styles.component}>
         <div className={styles.label}>{label}</div>
-        <div className={styles.radiosContainer}>
+        <div className={radiosContainerStyles}>
           {items.map(({ key, ...item }: RadioProps) => (
             <Radio
               skin={RadioSkin}