+++ source/renderer/app/containers/wallet/WalletAddPage.js
@@ -2,19 +2,22 @@
 import React, { Component } from 'react';
 import { inject, observer } from 'mobx-react';
 import WalletAdd from '../../components/wallet/WalletAdd';
-import WalletRestoreDialog from '../../components/wallet/WalletRestoreDialog';
 import WalletFileImportDialog from '../../components/wallet/file-import/WalletFileImportDialog';
 import WalletBackupDialog from '../../components/wallet/WalletBackupDialog';
 import WalletFileImportDialogContainer from './dialogs/WalletFileImportDialogContainer';
-import WalletRestoreDialogContainer from './dialogs/WalletRestoreDialogContainer';
 import WalletBackupDialogContainer from './dialogs/WalletBackupDialogContainer';
 import WalletCreateDialogContainer from './dialogs/WalletCreateDialogContainer';
+import WalletRestoreDialogContainer from './dialogs/WalletRestoreDialogContainer';
 import Layout from '../MainLayout';
 import type { InjectedProps } from '../../types/injectedPropsType';
 
 // TODO: Remove once the new wallet creation process is ready
 import WalletCreateDialogContainerOld from './dialogs/WalletCreateDialogContainerOld';
 import WalletCreateDialog from '../../components/wallet/WalletCreateDialog';
+
+// TODO: Remove once the new wallet restoration process is ready
+import WalletRestoreDialogContainerOld from './dialogs/WalletRestoreDialogContainerOld';
+import WalletRestoreDialog from '../../components/wallet/WalletRestoreDialog';
 
 type Props = InjectedProps;
 
@@ -30,16 +33,26 @@ export default class WalletAddPage extends Component<Props> {
   render() {
     const { actions, stores } = this.props;
     const { wallets, uiDialogs, app, networkStatus } = stores;
-    const { createWalletStep, useNewWalletCreationProcess } = wallets;
+    const {
+      createWalletStep,
+      createWalletUseNewProcess,
+      restoreWalletStep,
+      restoreWalletUseNewProcess,
+    } = wallets;
     const { isIncentivizedTestnet } = networkStatus;
     const {
       environment: { isMainnet, isTestnet },
     } = app;
 
-    const onWalletAdd = useNewWalletCreationProcess
+    const onCreateWallet = createWalletUseNewProcess
       ? () => actions.wallets.createWalletBegin.trigger()
       : // TODO: Remove once the new wallet creation process is ready
         () => actions.dialogs.open.trigger({ dialog: WalletCreateDialog });
+
+    const onRestoreWallet = restoreWalletUseNewProcess
+      ? () => actions.wallets.restoreWalletBegin.trigger()
+      : // TODO: Remove once the new wallet restoration process is ready
+        () => actions.dialogs.open.trigger({ dialog: WalletRestoreDialog });
 
     let content = null;
 
@@ -52,6 +65,8 @@ export default class WalletAddPage extends Component<Props> {
     } else if (uiDialogs.isOpen(WalletBackupDialog)) {
       content = <WalletBackupDialogContainer onClose={this.onClose} />;
     } else if (uiDialogs.isOpen(WalletRestoreDialog)) {
+      content = <WalletRestoreDialogContainerOld onClose={this.onClose} />;
+    } else if (restoreWalletStep !== null) {
       content = <WalletRestoreDialogContainer onClose={this.onClose} />;
     } else if (uiDialogs.isOpen(WalletFileImportDialog)) {
       content = <WalletFileImportDialogContainer onClose={this.onClose} />;
@@ -60,10 +75,8 @@ export default class WalletAddPage extends Component<Props> {
         <WalletAdd
           isMainnet={isMainnet}
           isTestnet={isTestnet}
-          onCreate={onWalletAdd}
-          onRestore={() =>
-            actions.dialogs.open.trigger({ dialog: WalletRestoreDialog })
-          }
+          onCreate={onCreateWallet}
+          onRestore={onRestoreWallet}
           onImportFile={() =>
             actions.dialogs.open.trigger({ dialog: WalletFileImportDialog })
           }