+++ source/renderer/app/themes/daedalus/dark-cardano.js
@@ -808,6 +808,8 @@ export const DARK_CARDANO_THEME_OUTPUT = {
     '--theme-wallet-restore-dialog-new-label-background-color':
       'rgba(255, 255, 255, 0.1)',
     '--theme-wallet-restore-dialog-new-label-color': '#ffffff',
+    '--theme-wallet-restore-dialog-step-walletType-hardwareWalletDisclaimer-text-color':
+      '#ea4c5b',
   },
 };